add all components for 'timestamp'
